refactor : Sections 객체 분리
