+++ src/main/java/nextstep/subway/line/domain/Line.java
@@ -17,9 +17,8 @@ public class Line extends BaseEntity {
     @Column(unique = true)
     private String name;
     private String color;
-
-    @OneToMany(mappedBy = "line", cascade = {CascadeType.PERSIST, CascadeType.MERGE}, orphanRemoval = true)
-    private List<Section> sections = new ArrayList<>();
+    @Embedded
+    private Sections sections = new Sections();
 
     public Line() {
     }
@@ -58,16 +57,7 @@ public class Line extends BaseEntity {
     }
 
     public List<Section> getSections() {
-        List<Section> sortedSections = new ArrayList<>();
-        Section section = findFirstSection();
-        sortedSections.add(section);
-
-        while (sortedSections.size() < this.sections.size()) {
-            section = findNextStation(section.getDownStation());
-            sortedSections.add(section);
-        }
-
-        return Collections.unmodifiableList(sortedSections);
+        return sections.getSections();
     }
 
     public List<Station> getStations() {
@@ -84,28 +74,6 @@ public class Line extends BaseEntity {
 
     private List<Station> getDownStations(List<Section> sortedSections) {
         return sortedSections.stream()
-                .map(Section::getDownStation)
-                .collect(Collectors.toList());
-    }
-
-    public Section findFirstSection() {
-        List<Station> downStations = getDownStations();
-
-        return sections.stream()
-                .filter(section -> !downStations.contains(section.getUpStation()))
-                .findFirst()
-                .orElseThrow(() -> new RuntimeException("상행 종점역을 찾을 수 없습니다."));
-    }
-
-    private Section findNextStation(Station downStation) {
-        return sections.stream()
-                .filter(section -> section.getUpStation().equals(downStation))
-                .findFirst()
-                .orElseThrow(() -> new RuntimeException("이어지는 구간을 찾을 수 없습니다."));
-    }
-
-    public List<Station> getDownStations() {
-        return sections.stream()
                 .map(Section::getDownStation)
                 .collect(Collectors.toList());
     }
