+++ src/main/java/nextstep/subway/station/domain/Station.java
@@ -30,8 +30,12 @@ public class Station extends BaseEntity {
 
     @Override
     public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
+        if (this == o) {
+            return true;
+        }
+        if (o == null || getClass() != o.getClass()) {
+            return false;
+        }
         Station station = (Station) o;
         return Objects.equals(id, station.id) &&
                 Objects.equals(name, station.name);
